modified the files which were causing issues
