+++ stubs/src/repo/UploadPage.js
@@ -2,7 +2,7 @@ let uploadPageJS = {};
 
 uploadPageJS.path = "src/repo";
 uploadPageJS.filename = 'UploadPage.js';
-uploadPageJS.content = `import uploadXpath '../xpaths/UploadPhoto.json'
+uploadPageJS.content = `import uploadXpath from '../xpaths/UploadPhoto.json'
 
 class UploadPage {
     image () {
