+++ stubs/src/tests/auth/LogoutPage.js
@@ -9,7 +9,8 @@ import Logout from "../../commons/Logout";
 describe('Logout the page', function () {
     before(() => {
         const emailValue = 'faizz.af@gmail.com';
-        const passwordValue = '123456';
+		const passwordValue = '123456';
+		Login.open()
         Login.login(emailValue,passwordValue)
     })
     it('should redirect to home page when logged out', () => {
