+++ stubs/src/tests/profile/UploadImage.js
@@ -11,7 +11,7 @@ describe("Do upload image in builderx page", function () {
 	});
 
 	it("should upload the image", () => {
-		const fileName ='/Users/biswajeet/Downloads/voting-7.jpg';
+		const fileName = "/path/to/image/file.jpg";
 		Login.login("faizz.af@gmail.com", "123456");
 		UploadPage.image().click();
 		UploadPage.settings().click();
